Add files via upload
This is the test model of the Docreader which comes with an option to change the gender of the voice engine depending on the user's selection. As in if the user selects on the radio button with the male label, then the voice engine initializes the voice id as male(incase you're using windows it will use mark). The initialization takes place once the button is clicked by the user.